Renaming service module to JokesApiService. Refactoring votes controller and fixing issues. All tests going green
