+++ app/controllers/api/v1/jokes_controller.rb
@@ -1,6 +1,6 @@
 class Api::V1::JokesController < ApplicationController
   def index
-    jokes = ReqResService.get_index_of_jokes
+    jokes = JokesApiService.get_index_of_jokes
     render json: { jokes: jokes }
   end
 end
